Move database to utils
* Database related interfaces and functions are moved to utils/Database.tsx. So that they can be used in other components too.
