+++ screens/Vocabulary.tsx
@@ -5,29 +5,24 @@ import { NavigationTabScreenProps } from 'react-navigation-tabs';
 import MaterialIcons from 'react-native-vector-icons/MaterialIcons';
 import Card from '../components/Card';
 
-import German from '../assets/languages/german.json';
-import Spanish from '../assets/languages/spanish.json';
-import French from '../assets/languages/french.json';
-import Russian from '../assets/languages/russian.json';
-
-export interface Word {
-  word: string;
-  translation: string;
-}
-
-export interface WordItem {
-  key: string;
-  word: string;
-  translation: string;
-}
+import { Word, WordItem, getDatabase } from '../utils/Database';
 
 export default function Vocabulary({ navigation }: NavigationTabScreenProps): JSX.Element {
   const [language, setLanguage] = useState('German');
-  const [database, setDatabase] = useState<WordItem[]>([]);
   const getSelectedLanguage = (): string => {
     const getLanguage = navigation.getParam('getLanguage') as () => string;
     return getLanguage();
   };
+
+  const [database, setDatabase] = useState<WordItem[]>([]);
+  const getWords = () => {
+    if (database.length > 0) return database;
+
+    const newDatabase = getDatabase(language, database);
+    setDatabase(newDatabase);
+    return newDatabase;
+  };
+
   useEffect(() => {
     navigation.addListener('didFocus', () => {
       const selectedLanguage = getSelectedLanguage();
@@ -38,28 +33,6 @@ export default function Vocabulary({ navigation }: NavigationTabScreenProps): JS
       }
     });
   });
-  const getDatabase = () => {
-    if (database.length > 0) return database;
-
-    let loadedDatabase: Word[] = German as Word[];
-    if (language == 'Spanish') loadedDatabase = Spanish as Word[];
-    if (language == 'French') loadedDatabase = French as Word[];
-    if (language == 'Russian') loadedDatabase = Russian as Word[];
-
-    const languageDatabase: WordItem[] = [];
-    let key = 1;
-    loadedDatabase.forEach((element: Word) => {
-      const newWord: WordItem = {
-        key: key.toString(),
-        word: element.word,
-        translation: element.translation,
-      };
-      key = key + 1;
-      languageDatabase.push(newWord);
-    });
-    setDatabase(languageDatabase);
-    return database;
-  };
 
   const [modalOpen, setModalOpen] = useState(false);
   const [item, setItem] = useState<Word>({ word: '', translation: '' });
@@ -98,7 +71,7 @@ export default function Vocabulary({ navigation }: NavigationTabScreenProps): JS
       </Modal>
 
       <FlatList
-        data={getDatabase()}
+        data={getWords()}
         renderItem={({ item }) => (
           <TouchableOpacity onPress={() => openModal(item)}>
             <Card>
